[disasm] Update disassembler sources to compile with Clang on MacOS
Details

  On MacOS, there were several old constructs that no longer build
  correctly.  They are remedied in a way that renders no semantic
  changes to the code, but allows the code to build on Linux and MacOS.

Testing

  Now, on MacOS:

    disasm ./system/Texts.Obj

  successfully disassembles the Texts module.

+++ disasm/objio.cpp
@@ -111,7 +111,7 @@ namespace objio
         int i, len;
 
         read_num(len);
-        str = new unsigned char[len + 1];
+        str = new unsigned char[static_cast<unsigned long>(len + 1)];
         for (i = 0; i < len; ++i)
             read_char(str[i]);
         str[i] = '\0';
